Refactor so that pure pursuit controller is more self contained
