+++ commons/src/main/scala/com/lynbrookrobotics/potassium/commons/drivetrain/UnicycleDrive.scala
@@ -5,7 +5,7 @@ import com.lynbrookrobotics.potassium.{PeriodicSignal, Signal, SignalLike}
 import squants.motion.AngularVelocity
 import squants.{Acceleration, Angle, Dimensionless, Each, Length, Percent, Velocity}
 import com.lynbrookrobotics.potassium.commons.drivetrain.UnicycleMotionProfileControllers
-import com.lynbrookrobotics.potassium.commons.position.AdvancedPositionTasks
+import com.lynbrookrobotics.potassium.commons.position.PurePursuitTasks
 import squants.space.Feet
 import squants.time.Milliseconds
 
@@ -67,7 +67,7 @@ trait UnicycleDrive extends Drive { self =>
     */
   protected def convertUnicycleToDrive(uni: UnicycleSignal): DriveSignal
   
-  object UnicycleControllers extends AdvancedPositionControllers with UnicycleMotionProfileControllers with UnicycleCoreControllers{
+  object UnicycleControllers extends PurePursuitControllers with UnicycleMotionProfileControllers with UnicycleCoreControllers{
     type DriveSignal = self.DriveSignal
     type DrivetrainHardware = self.Hardware
     type DrivetrainProperties = self.Properties
@@ -94,7 +94,7 @@ trait UnicycleDrive extends Drive { self =>
 
   import UnicycleControllers._
 
-  object unicycleTasks extends UnicycleCoreTasks with AdvancedPositionTasks {
+  object unicycleTasks extends UnicycleCoreTasks with PurePursuitTasks {
     type Drivetrain = self.Drivetrain
     override val controllers = UnicycleControllers
   }
